refactor: extract shared condition evaluation into ConditionEvaluator utility
Extracts the duplicated 12-condition switch from both
_item-action-card-execution.mjs_ (instance method) and
_damage-processor.mjs_ (static method) into a shared
ConditionEvaluator class. Both callers now delegate to
ConditionEvaluator.evaluate(). No behavioral change.

Adding a new condition type is now a 1-place change instead of 2.
