+++ module/documents/mixins/item-action-card-execution.mjs
@@ -2,7 +2,7 @@ import { Logger } from "../../services/logger.mjs";
 import { ResourceValidator } from "../../services/resource-validator.mjs";
 import { TransformationApplicator } from "../../services/transformation-applicator.mjs";
 import { RepetitionHandler } from "../../services/repetition-handler.mjs";
-import { ERPSRollUtilities } from "../../utils/roll-utilities.mjs";
+import { ConditionEvaluator } from "../../helpers/condition-evaluator.mjs";
 import {
   StatusEffectApplicator,
   TargetResolver,
@@ -465,50 +465,16 @@ export function ItemActionCardExecutionMixin(Base) {
       actor = null,
       formula = null,
     ) {
-      switch (condition) {
-        case "never":
-          return false;
-        case "oneSuccess":
-          return oneHit;
-        case "twoSuccesses":
-          return bothHit;
-        case "rollValue":
-          return rollTotal >= threshold;
-        case "rollUnderValue":
-          return rollTotal < threshold;
-        case "rollEven":
-          return rollTotal % 2 === 0;
-        case "rollOdd":
-          return rollTotal % 2 !== 0;
-        case "rollOnValue":
-          return rollTotal === threshold;
-        case "zeroSuccesses":
-          return !oneHit;
-        case "always":
-          return true;
-        case "criticalSuccess":
-        case "criticalFailure": {
-          if (!roll || !actor || !formula) {
-            return false;
-          }
-
-          const actorRollData = actor.getRollData();
-          const criticalStates = ERPSRollUtilities.determineCriticalStates({
-            roll,
-            thresholds: actorRollData.hiddenAbilities,
-            formula,
-            critAllowed: true,
-          });
-          
-          const result = condition === "criticalSuccess"
-            ? criticalStates.critHit && !criticalStates.stolenCrit
-            : criticalStates.critMiss && !criticalStates.savedMiss;
-
-          return result;
-        }
-        default:
-          return false;
-      }
+      return ConditionEvaluator.evaluate(
+        condition,
+        oneHit,
+        bothHit,
+        rollTotal,
+        threshold,
+        roll,
+        actor,
+        formula,
+      );
     }
 
     /**
